Split WindowEventBridge into four focused dispatchers (#373)
