+++ minimark/Views/Window/Coordination/WindowEventBridge.swift
@@ -3,123 +3,56 @@ import Foundation
 
 /// Translates SwiftUI window events (.background WindowAccessor, .onAppear,
 /// .onDisappear, .onChange of various properties) into mutations on the
-/// extracted controllers. Holds no business logic of its own — every method
-/// is a thin reducer dispatching to one or more collaborators.
-///
-/// Bundled here because every method is the same shape: a SwiftUI hook fires,
-/// the bridge decides which collaborators need to know. Splitting one method
-/// per file would just disperse that uniform pattern.
+/// extracted controllers. A thin composite over four focused dispatchers —
+/// each dispatcher owns its own slice of deps (see #373).
 @MainActor
 final class WindowEventBridge {
-    let openDocumentPathTracker = OpenDocumentPathTracker()
-
-    private let shell: WindowShellController
-    private let folderWatchOpen: WindowFolderWatchOpenController
-    private let sidebarDocumentController: SidebarDocumentController
-    private let settingsStore: SettingsStore
-    private let groupStateControllerProvider: () -> SidebarGroupStateController?
-    private let favoriteWorkspaceControllerProvider: () -> FavoriteWorkspaceController?
-    private let appearanceControllerProvider: () -> WindowAppearanceController?
-    private let uiTestLaunchCoordinatorProvider: () -> UITestLaunchCoordinator?
-    private let refreshWindowShellState: () -> Void
+    let hostLifecycle: WindowHostLifecycleDispatcher
+    let documentSync: WindowDocumentSyncDispatcher
+    let favoriteWorkspace: FavoriteWorkspaceEventDispatcher
+    let groupState: GroupStateEventDispatcher
 
     init(
-        shell: WindowShellController,
-        folderWatchOpen: WindowFolderWatchOpenController,
-        sidebarDocumentController: SidebarDocumentController,
-        settingsStore: SettingsStore,
-        groupStateControllerProvider: @escaping () -> SidebarGroupStateController?,
-        favoriteWorkspaceControllerProvider: @escaping () -> FavoriteWorkspaceController?,
-        appearanceControllerProvider: @escaping () -> WindowAppearanceController?,
-        uiTestLaunchCoordinatorProvider: @escaping () -> UITestLaunchCoordinator?,
-        refreshWindowShellState: @escaping () -> Void
+        hostLifecycle: WindowHostLifecycleDispatcher,
+        documentSync: WindowDocumentSyncDispatcher,
+        favoriteWorkspace: FavoriteWorkspaceEventDispatcher,
+        groupState: GroupStateEventDispatcher
     ) {
-        self.shell = shell
-        self.folderWatchOpen = folderWatchOpen
-        self.sidebarDocumentController = sidebarDocumentController
-        self.settingsStore = settingsStore
-        self.groupStateControllerProvider = groupStateControllerProvider
-        self.favoriteWorkspaceControllerProvider = favoriteWorkspaceControllerProvider
-        self.appearanceControllerProvider = appearanceControllerProvider
-        self.uiTestLaunchCoordinatorProvider = uiTestLaunchCoordinatorProvider
-        self.refreshWindowShellState = refreshWindowShellState
+        self.hostLifecycle = hostLifecycle
+        self.documentSync = documentSync
+        self.favoriteWorkspace = favoriteWorkspace
+        self.groupState = groupState
     }
+
+    /// Read-through surface kept for `WindowRootView` consumers. The tracker
+    /// itself is owned by `documentSync`, which is the only dispatcher that
+    /// mutates it.
+    var openDocumentPathTracker: OpenDocumentPathTracker { documentSync.openDocumentPathTracker }
 
     func handleWindowAccessorUpdate(_ window: NSWindow?) {
-        guard shell.updateHostWindow(window) else { return }
-        handleHostWindowChange()
-    }
-
-    private func handleHostWindowChange() {
-        refreshWindowShellState()
-        uiTestLaunchCoordinatorProvider()?.applyConfigurationIfNeeded()
-        if shell.hostWindow != nil, folderWatchOpen.hasPendingEvents {
-            folderWatchOpen.flush()
-        }
+        hostLifecycle.handleWindowAccessorUpdate(window)
     }
 
     func handleWindowAppear() {
-        let groupState = groupStateControllerProvider()
-        groupState?.configureSortModes(
-            sortMode: settingsStore.currentSettings.sidebarGroupSortMode,
-            fileSortMode: settingsStore.currentSettings.sidebarSortMode
-        )
-        groupState?.updateDocuments(
-            sidebarDocumentController.documents,
-            rowStates: sidebarDocumentController.rowStates
-        )
-        groupState?.observeRowStates(from: sidebarDocumentController)
-        openDocumentPathTracker.update(from: sidebarDocumentController.documents)
-        shell.configureDockTile()
+        documentSync.handleWindowAppear()
     }
 
     func handleWindowDisappear() {
-        shell.clearDockTile()
+        documentSync.handleWindowDisappear()
     }
 
     func handleDocumentListChange() {
-        groupStateControllerProvider()?.updateDocuments(
-            sidebarDocumentController.documents,
-            rowStates: sidebarDocumentController.rowStates
-        )
-        openDocumentPathTracker.update(from: sidebarDocumentController.documents)
+        documentSync.handleDocumentListChange()
     }
 
     func handleFavoriteWorkspaceStateChange(_ newState: FavoriteWorkspaceState?) {
-        guard let favoriteID = favoriteWorkspaceControllerProvider()?.activeFavoriteID,
-              var state = newState else { return }
-        state.lockedAppearance = appearanceControllerProvider()?.lockedAppearance
-        settingsStore.updateFavoriteWorkspaceState(id: favoriteID, workspaceState: state)
+        favoriteWorkspace.handleFavoriteWorkspaceStateChange(newState)
     }
 
     func handleGroupStateChange(
         oldSnapshot: SidebarGroupStateController.WorkspaceStateSnapshot,
         newSnapshot: SidebarGroupStateController.WorkspaceStateSnapshot
     ) {
-        if let favoriteController = favoriteWorkspaceControllerProvider(), favoriteController.isActive {
-            let needsUpdate =
-                favoriteController.activeFavoriteWorkspaceState?.pinnedGroupIDs != newSnapshot.pinnedGroupIDs ||
-                favoriteController.activeFavoriteWorkspaceState?.collapsedGroupIDs != newSnapshot.collapsedGroupIDs ||
-                favoriteController.activeFavoriteWorkspaceState?.groupSortMode != newSnapshot.sortMode ||
-                favoriteController.activeFavoriteWorkspaceState?.fileSortMode != newSnapshot.fileSortMode ||
-                favoriteController.activeFavoriteWorkspaceState?.manualGroupOrder != newSnapshot.manualGroupOrder
-
-            if needsUpdate {
-                favoriteController.updateGroupState(
-                    pinnedGroupIDs: newSnapshot.pinnedGroupIDs,
-                    collapsedGroupIDs: newSnapshot.collapsedGroupIDs,
-                    groupSortMode: newSnapshot.sortMode,
-                    fileSortMode: newSnapshot.fileSortMode,
-                    manualGroupOrder: newSnapshot.manualGroupOrder
-                )
-            }
-        } else {
-            if oldSnapshot.sortMode != newSnapshot.sortMode {
-                settingsStore.updateSidebarGroupSortMode(newSnapshot.sortMode)
-            }
-            if oldSnapshot.fileSortMode != newSnapshot.fileSortMode {
-                settingsStore.updateSidebarSortMode(newSnapshot.fileSortMode)
-            }
-        }
+        groupState.handleGroupStateChange(oldSnapshot: oldSnapshot, newSnapshot: newSnapshot)
     }
 }
